Move utils in new common module
