+++ snips_nlu/entity_parser/entity_parser.py
@@ -6,14 +6,14 @@ from abc import ABCMeta, abstractmethod
 from future.builtins import object
 from future.utils import with_metaclass
 
-from snips_nlu.utils import LimitedSizeDict
+from snips_nlu.common.dict_utils import LimitedSizeDict
 
 # pylint: disable=ungrouped-imports
 
 try:
     from abc import abstractclassmethod
 except ImportError:
-    from snips_nlu.utils import abstractclassmethod
+    from snips_nlu.common.abc_utils import abstractclassmethod
 
 
 # pylint: enable=ungrouped-imports
