+++ snips_nlu/slot_filler/features_utils.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 from snips_nlu_utils import compute_all_ngrams
 
 from snips_nlu.constants import (END, RES_MATCH_RANGE, START)
-from snips_nlu.utils import LimitedSizeDict
+from snips_nlu.common.dict_utils import LimitedSizeDict
 
 _NGRAMS_CACHE = LimitedSizeDict(size_limit=1000)
