+++ snips_nlu/slot_filler/slot_filler.py
@@ -1,4 +1,4 @@
-from abc import ABCMeta, abstractmethod
+from abc import abstractmethod, ABCMeta
 
 from future.utils import with_metaclass
